+++ qupath-gui-fx/src/main/java/qupath/lib/gui/commands/SummaryMeasurementTableCommand.java
@@ -66,6 +66,7 @@ import javafx.scene.control.SelectionMode;
 import javafx.scene.control.SplitPane;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableRow;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TableView.TableViewSelectionModel;
 import javafx.scene.control.TextField;
@@ -73,6 +74,8 @@ import javafx.scene.control.ToggleButton;
 import javafx.scene.image.Image;
 import javafx.scene.input.Clipboard;
 import javafx.scene.input.ClipboardContent;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyCodeCombination;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Pane;
@@ -185,6 +188,21 @@ public class SummaryMeasurementTableCommand {
 		else
 			name = PathObjectTools.getSuitableName(type, false) + " results - " + displayedName;
 
+		// Handle double-click as a way to center on a ROI
+//		var enter = new KeyCodeCombination(KeyCode.ENTER);
+		table.setRowFactory(params -> {
+			var row = new TableRow<PathObject>() ;
+			row.setOnMouseClicked(e -> {
+				if (e.getClickCount() == 2) {
+					maybeCenterROI(row.getItem());
+				}
+			});
+//			row.setOnKeyPressed(e -> {
+//				if (enter.match(e))
+//					maybeCenterROI(row.getItem());
+//			});
+			return row;
+		});
 
 		// Create columns according to the table model
 //		for (int i = 0; i < model.getColumnCount(); i++) {
@@ -473,6 +491,15 @@ public class SummaryMeasurementTableCommand {
 
 	}
 
+	
+	private void maybeCenterROI(PathObject pathObject) {
+		if (pathObject == null)
+			return;
+		var roi = pathObject.getROI();
+		var viewer = qupath.getViewer();
+		if (roi != null && viewer != null && viewer.getHierarchy() != null)
+			viewer.centerROI(roi);
+	}
 
 
 
@@ -597,8 +624,9 @@ public class SummaryMeasurementTableCommand {
 
 
 				setOnMouseClicked(e -> {
-					if (e.getClickCount() > 1 && histogramDisplay != null) {
+					if (e.isAltDown() && histogramDisplay != null) {
 						histogramDisplay.showHistogram(getTableColumn().getText());
+						e.consume();
 						//	            		showChart(column);
 					}
 				});